2007-09-10  Dan Williams  <dcbw@redhat.com>
	* NetworkManagerAP.c
	  NetworkManagerAP.h
	  NetworkManagerPolicy.c
	  NetworkManagerSystem.c
	  NetworkManagerUtils.c
	  NetworkManagerUtils.h
	  nm-device-802-11-wireless.c
	  nm-device-802-3-ethernet.c
	  nm-hal-manager.c
	  nm-manager.c
	  vpn-manager/nm-dbus-vpn.c
		- Warning fixes; casts and removal of unused variables



git-svn-id: http://svn-archive.gnome.org/svn/NetworkManager/trunk@2779 4912f4e0-d625-0410-9fb7-b9a5a253dbdc

+++ src/NetworkManagerAP.h
@@ -129,7 +129,7 @@ void				nm_ap_print_self (NMAccessPoint *ap, const char * prefix);
  */
 gboolean			nm_ap_has_manufacturer_default_ssid	(NMAccessPoint *ap);
 
-guint32 freq_to_channel (double freq);
-double  channel_to_freq (guint32 channel);
+guint32 freq_to_channel (guint32 freq);
+guint32  channel_to_freq (guint32 channel);
 
 #endif /* NM_ACCESS_POINT_H */
